Polish unit test tests/unit_tests/distributed/mfsdp_v2/test_pipeline_parallel.py
